Change float to double

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/CompetitionCodev1.java
@@ -1,5 +1,4 @@
 package org.firstinspires.ftc.teamcode;
-import static org.firstinspires.ftc.robotcore.external.BlocksOpModeCompanion.hardwareMap;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.DcMotor;
@@ -12,14 +11,13 @@ public class CompetitionCodev1 extends LinearOpMode {
     private DcMotor backLeft;
     private DcMotor frontRight;
     private DcMotor backRight;
-
     /**
      * Basic mecanum driving
      */
     private void mecanum_drive() {
-        float y;
+        double y;
         double x;
-        float rx;
+        double rx;
         double denominator;
 
         // Y and X are combined to make a fraction that creates the power values for the motors. RX is for rotating the robot and only applies to the right stick
